pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arkiv"
version = "0.1.0"
description = "Universal personal data format. JSONL in, SQL out, MCP to LLMs."
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
mcp = ["mcp[cli]"]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "mcp[cli]",
]

[project.scripts]
arkiv = "arkiv.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 88
target-version = ["py38"]
